Change command to run codemod in README
It appears since the project was converted to use codemod-cli you can no longer use it by linking to the raw js file in github. Also as it's not an npm package you can't install it globally. The only way I could get it running was to use the following command:

```
npx github:rondale-sc/es5-getter-ember-codemod es5-getter-ember-codemod app/**/*.js
```

Also as jscodeshift is now part of codemod-cli you no longer need to install it yourself.