Update codemod-cli
This apparently brings a more up-to-date parser, especially fixes #24.

Had to update a bunch of test fixtures, as they previously contained invalid syntax (duplicate variable declarations), which the new parser did not accept. Also aligned JS and TS fixtures, as the differences were not related to TS at all (wrapped code in a native class), so this was a bit confusing for me.